ArC: convert ComponentsProviderGenerator to Gizmo 2
This includes converting the Gizmo 1 portion of `Types` to Gizmo 2.
The API is changed somewhat; the new converter of build time Jandex
`Type`s to runtime reflective `Type`s using Gizmo 2 is available
as `RuntimeTypeCreator`. It has a single factory method `of()`,
a few methods `with*()` to set options, and a single method
to generate the bytecode: `create()`.

BREAKING CHANGE: The following `ContextConfigurator` method

```
ContextConfigurator creator(Function<MethodCreator, ResultHandle> creator)
```

was changed to

```
ContextConfigurator creator(BiFunction<BlockCreator, ParamVar, Expr> creator)
```